Module amazon_sp_vendor_direct_fulfillment_sandbox_test_data::apis::vendor_df_sandboxtransactionstatus_api

Enums

GetOrderScenariosError

struct for typed errors of method get_order_scenarios

Functions

get_order_scenarios

Returns the status of the transaction indicated by the specified transactionId. If the transaction was successful, also returns the requested test order data.